Refactor - ResultView의 Rank 계산 로직 분리
